pyproject.toml
[project]
name = "devops-engineer-from-scratch-project-313"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.0",
    "fastapi-cli>=0.0.16",
    "psycopg[binary]>=3.3.0",
    "python-dotenv>=1.2.1",
    "sentry-sdk>=2.46.0",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "ruff>=0.14.7",
]
